hotstuff/requirements.txt
grpcio
grpcio-tools
protobuf
betterproto[compiler]